refactor(article): リポジトリモジュールをサービスに統合しコード構造を整理
repositoryモジュールを削除し、その機能をserviceモジュールに統合
公開APIの再エクスポートを整理し、不要なコメントを削除
