+++ src/core/article/mod.rs
@@ -1,5 +1,4 @@
 pub mod model;
-pub mod repository;
 pub mod service;
 
 // 公開APIの再エクスポート
@@ -11,11 +10,10 @@ pub use model::{
     Article, ArticleMetadata, ArticleStatus,
 };
 
-// repository.rsから
-pub use repository::{
-    search_article_contents, search_articles, search_backlog_articles_light, store_article_content,
-    ArticleContent, ArticleContentQuery, ArticleQuery,
+// repository.rsから（統合後）
+pub use service::{
+    fetch_and_store_article, fetch_and_store_article_with_client, get_article_content,
+    get_article_content_with_client, search_article_contents, search_articles,
+    search_backlog_articles_light, store_article_content, ArticleContent, ArticleContentQuery,
+    ArticleQuery,
 };
-
-// service.rsから
-pub use service::{get_article_content, get_article_content_with_client};
